+++ drivers/common/openserial.h
@@ -118,7 +118,7 @@ BEGIN_PACK
 typedef struct{
    uint16_t    track_instance;
    uint8_t     track_owner[8];
-   uint16_t    seqnum;
+   uint32_t    seqnum;
    uint8_t     l3Source[8];
    uint8_t     l3Dest[8];
 } evtPktData_t;
@@ -227,8 +227,8 @@ void  openserial_statRx(OpenQueueEntry_t* msg);
 void  openserial_statTx(OpenQueueEntry_t* msg);
 void  openserial_statPktTimeout(OpenQueueEntry_t* msg);
 void  openserial_statPktBufferOverflow(OpenQueueEntry_t* msg);
-void  openserial_statDataGen(uint16_t seqnum, track_t *track, open_addr_t *src, open_addr_t *dest);
-void  openserial_statDataRx(uint16_t seqnum, track_t *track, open_addr_t *src, open_addr_t *dest);
+void  openserial_statDataGen(uint32_t seqnum, track_t *track, open_addr_t *src, open_addr_t *dest);
+void  openserial_statDataRx(uint32_t seqnum, track_t *track, open_addr_t *src, open_addr_t *dest);
 void  openserial_statDIOtx(void);
 void  openserial_statDAOtx(uint8_t *parent);
